Sample cells from the simulations

generate_experiment() runs samples cells along the different simulations. experiment_snapshot() assumes that cells are sampled from a heterogeneous pool of cells. Cells will thus be sampled uniformily from the trajectory. experiment_synchronised() assumes that all the cells are synchronised and are sampled at different timepoints.

generate_experiment(model)

list_experiment_samplers()

experiment_snapshot(
  realcount = NULL,
  map_reference_cpm = TRUE,
  map_reference_ls = TRUE,
  weight_bw = 0.1
)

experiment_synchronised(
  realcount = NULL,
  map_reference_cpm = TRUE,
  map_reference_ls = TRUE,
  num_timepoints = 8,
  pct_between = 0.75
)

Arguments

model
A dyngen intermediary model for which the simulations have been run with generate_cells().

realcount
The name of a dataset in realcounts. If NULL, a random dataset will be sampled from realcounts.

map_reference_cpm
Whether or not to try to match the CPM distribution to that of a reference dataset.

map_reference_ls
Whether or not to try to match the distribution of the library sizes to that of the reference dataset.

weight_bw
[snapshot] A bandwidth parameter for determining the distribution of cells along each edge in order to perform weighted sampling.

num_timepoints
[synchronised] The number of time points used in the experiment.

pct_between
[synchronised] The percentage of 'unused' simulation time.

Value

A dyngen model.

Examples

names(list_experiment_samplers())

#> [1] "snapshot"     "synchronised"

model <- 
  initialise_model(
    backbone = backbone_bifurcating(),
    experiment = experiment_synchronised()
  )
# \dontshow{
# actually use a smaller example 
# to reduce execution time during
# testing of the examples
model <- initialise_model(
  backbone = model$backbone,
  num_cells = 5,
  num_targets = 0,
  num_hks = 0,
  gold_standard_params = gold_standard_default(census_interval = 1, tau = 0.1),
  simulation_params = simulation_default(
    burn_time = 10,
    total_time = 10,
    census_interval = 1,
    ssa_algorithm = ssa_etl(tau = 0.1),
    experiment_params = simulation_type_wild_type(num_simulations = 1)
  )
)
# }
# \donttest{
model <- 
  model %>%
  generate_tf_network() %>%
  generate_feature_network() %>%
  generate_kinetics() %>%
  generate_gold_standard() %>%
  generate_cells() %>%
  generate_experiment() 

#> Generating TF network
#> Sampling feature network from real network
#> Generating kinetics for 36 features
#> Generating formulae
#> Generating gold standard mod changes
#> Precompiling reactions for gold standard
#> Running gold simulations
#>   |                                                  | 0 % elapsed=00s     |========                                          | 14% elapsed=00s, remaining~00s  |===============                                   | 29% elapsed=00s, remaining~00s  |======================                            | 43% elapsed=00s, remaining~00s  |=============================                     | 57% elapsed=00s, remaining~00s  |====================================              | 71% elapsed=00s, remaining~00s  |===========================================       | 86% elapsed=00s, remaining~00s  |==================================================| 100% elapsed=00s, remaining~00s
#> Precompiling reactions for simulations
#> Running 1 simulations
#> Mapping simulations to gold standard

#> Warning: Simulation does not contain all gold standard edges. This simulation likely suffers from bad kinetics; choose a different seed and rerun.

#> Performing dimred
#> Simulating experiment

#> Error in rmultinom(1, lib_size, gene_vals): invalid second argument 'size'

  
dataset <- as_dyno(model)

#> Error in as(log2(counts + 1), "dgCMatrix"): no method or default for coercing “numeric” to “dgCMatrix”

# }