Determine the kinetics of the feature network

generate_kinetics() samples the kinetics of genes in the feature network for which the kinetics have not yet been defined. kinetics_default() is used to configure parameters pertaining this process. kinetics_random_distributions() will do the same, but the distributions are also randomised.

generate_kinetics(model)

kinetics_default()

kinetics_random_distributions()

Arguments

model

A dyngen intermediary model for which the feature network has been generated with generate_feature_network().

Value

A dyngen model.

Details

To write different kinetics settings, you need to write three functions with interface function(feature_info, feature_network, cache_dir, verbose). Described below are the default kinetics samplers.

sampler_tfs() mutates the feature_info data frame by adding the following columns:

transcription_rate: the rate at which pre-mRNAs are transcribed, in pre-mRNA / hour. Default distribution: U(1, 2).

translation_rate: the rate at which mRNAs are translated into proteins, in protein / mRNA / hour. Default distribution: U(100, 150).

mrna_halflife: the half-life of (pre-)mRNA molecules, in hours. Default distribution: U(2.5, 5).

protein_halflife: the half-life of proteins, in hours. Default distribution: U(5, 10).

splicing_rate: the rate at which pre-mRNAs are spliced into mRNAs, in reactions / hour. Default value: log(2) / (10/60), which corresponds to a half-life of 10 minutes.

independence: the degree to which all regulators need to be bound for transcription to occur (0), or whether transcription can occur if only one of the regulators is bound (1).

sampler_nontfs() samples the transcription_rate, translation_rate, mrna_halflife and protein_halflife from a supplementary file of Schwannhäusser et al., 2011, doi.org/10.1038/nature10098. splicing_rate is by default the same as in sampler_tfs(). independence is sampled from U(0, 1).

sampler_interactions() mutates the feature_network data frame by adding the following columns.

effect: the effect of the interaction; upregulating = +1, downregulating = -1. By default, sampled from -1, 1 with probabilities .25, .75.

strength: the strength of the interaction. Default distribution: 10^U(0, 2).

hill: the hill coefficient. Default distribution: N(2, 2) with a minimum of 1 and a maximum of 10.

See also

Examples

model <- 
  initialise_model(
    backbone = backbone_bifurcating(),
    kinetics_params = kinetics_default()
  )

# \donttest{
data("example_model")
model <- example_model %>%
  generate_kinetics()
# }