List of all predefined backbone models

A module is a group of genes which, to some extent, shows the same expression behaviour. Several modules are connected together such that one or more genes from one module will regulate the expression of another module. By creating chains of modules, a dynamic behaviour in gene regulation can be created.

list_backbones()

backbone_bifurcating()

backbone_bifurcating_converging()

backbone_bifurcating_cycle()

backbone_bifurcating_loop()

backbone_branching(
  num_modifications = rbinom(1, size = 6, 0.25) + 1,
  min_degree = 3,
  max_degree = sample(min_degree:5, 1)
)

backbone_binary_tree(num_modifications = rbinom(1, size = 6, 0.25) + 1)

backbone_consecutive_bifurcating()

backbone_trifurcating()

backbone_converging()

backbone_cycle()

backbone_cycle_simple()

backbone_linear()

backbone_linear_simple()

backbone_disconnected(
  left_backbone = list_backbones() %>% keep(., names(.) != "disconnected") %>%
    sample(1) %>% first(),
  right_backbone = list_backbones() %>% keep(., names(.) != "disconnected") %>%
    sample(1) %>% first(),
  num_common_modules = 10
)

Arguments

num_modifications
The number of branch points in the generated backbone.

min_degree
The minimum degree of each node in the backbone.

max_degree
The maximum degree of each node in the backbone.

left_backbone
A backbone (other than a disconnected backbone), see list_backbones().

right_backbone
A backbone (other than a disconnected backbone), see list_backbones().

num_common_modules
The number of modules which are regulated by either backbone.

Value

A list of all the available backbone generators.

See also

Examples

names(list_backbones())

#>  [1] "bifurcating"             "bifurcating_converging" 
#>  [3] "bifurcating_cycle"       "bifurcating_loop"       
#>  [5] "binary_tree"             "branching"              
#>  [7] "consecutive_bifurcating" "converging"             
#>  [9] "cycle"                   "cycle_simple"           
#> [11] "disconnected"            "linear"                 
#> [13] "linear_simple"           "trifurcating"           

bb <- backbone_bifurcating()
bb <- backbone_bifurcating_converging()
bb <- backbone_bifurcating_cycle()
bb <- backbone_bifurcating_loop()
bb <- backbone_binary_tree()
bb <- backbone_branching()
bb <- backbone_consecutive_bifurcating()
bb <- backbone_converging()
bb <- backbone_cycle()
bb <- backbone_cycle_simple()
bb <- backbone_disconnected()
bb <- backbone_linear()
bb <- backbone_linear_simple()
bb <- backbone_trifurcating()

model <- initialise_model(
  backbone = bb
)
# \dontshow{
# actually use a smaller example 
# to reduce execution time during
# testing of the examples
model <- initialise_model(
  backbone = model$backbone,
  num_cells = 5,
  num_targets = 0,
  num_hks = 0,
  gold_standard_params = gold_standard_default(census_interval = 1, tau = 0.1),
  simulation_params = simulation_default(
    burn_time = 10,
    total_time = 10,
    census_interval = 1,
    ssa_algorithm = ssa_etl(tau = 0.1),
    experiment_params = simulation_type_wild_type(num_simulations = 1)
  )
)
# }
# \donttest{
model <- model %>%
  generate_tf_network() %>%
  generate_feature_network() %>%
  generate_kinetics() %>%
  generate_gold_standard() %>%
  generate_cells() %>%
  generate_experiment() 

#> Generating TF network
#> Sampling feature network from real network
#> Generating kinetics for 53 features
#> Generating formulae
#> Generating gold standard mod changes
#> Precompiling reactions for gold standard
#> Running gold simulations
#>   |                                                  | 0 % elapsed=00s     |=====                                             | 10% elapsed=00s, remaining~00s  |==========                                        | 20% elapsed=00s, remaining~00s  |===============                                   | 30% elapsed=00s, remaining~00s  |====================                              | 40% elapsed=00s, remaining~00s  |=========================                         | 50% elapsed=00s, remaining~00s  |==============================                    | 60% elapsed=00s, remaining~00s  |===================================               | 70% elapsed=00s, remaining~00s  |========================================          | 80% elapsed=00s, remaining~00s  |=============================================     | 90% elapsed=00s, remaining~00s  |==================================================| 100% elapsed=00s, remaining~00s
#> Precompiling reactions for simulations
#> Running 1 simulations
#> Mapping simulations to gold standard

#> Warning: Simulation does not contain all gold standard edges. This simulation likely suffers from bad kinetics; choose a different seed and rerun.

#> Performing dimred
#> Simulating experiment

#> Error in rmultinom(1, lib_size, gene_vals): invalid second argument 'size'

  
dataset <- as_dyno(model)

#> Error in as(log2(counts + 1), "dgCMatrix"): no method or default for coercing “numeric” to “dgCMatrix”

# }